backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.15.0
python-multipart==0.0.6


